feat(dsy-854-progress-indicator): changes suggested by the code review
