--- lib/components/NatLoading/NatLoading.styles.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components/native';
-
-export const LoadingWrapper = styled.View`
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  right: 0;
-  left: 0;
-  top: 0;
-  bottom: 0;
-  z-index: 10;
-`;
